WIP: Make the GC Adapter work on macOS
On macOS, the HID driver opens the GC adapter device when it’s plugged in,
which prevents Dolphin from opening it. However, if Dolphin tries to open the
device as soon as it’s plugged in, it will usually outrun the HID driver and
succeed.

This change simulate the device being unplugged and replugged by resetting it
if Dolphin fails to open it.

TODO: Make it retry a couple of times (it actually works first try on my
machine every time, but it may not be the case everywhere).

FIXME: This change requires a newer version of libusb; the current one does
not actually resets the device on macOS, and this can’t easily be done outside
of the library.